Make application endpoints more like interactive endpoints
